refactor: View를 Input과 Output으로 분리
